+++ libffmpegxx/avcodec/AVPacketImpl.cpp
@@ -6,7 +6,11 @@
 #include <stdexcept>
 
 extern "C" {
+#if LIBAVCODEC_VERSION_MAJOR >= 58 && LIBAVCODEC_VERSION_MINOR >=78 && LIBAVCODEC_VERSION_MICRO >= 100
 #include <libavcodec/packet.h>
+#else
+#include <libavcodec/avcodec.h>
+#endif
 #include <libavutil/avutil.h>
 }